backend/requirements.txt
fastapi
uvicorn
python-multipart
litellm
python-dotenv
sqlalchemy